Minor type hinting changes, because typing.AnyStr has changed slightly.
typing.AnyStr now is exclusive, so you can refer to Any String, but the
first context of it within a function's type definitions defines it for
all other subsequent implementations of it within that function. This
will result in undesired behavior when using imaplibext in IDEs such as
PyCharm, or when running code inspections which also test for data type
compatibility. We fix this by redefining AnyStr locally to be the type
inspection of Union[str, bytes] like the original AnyStr implementation
had been, effectively restoring the 'old ways' within the imaplibext
library itself instead of relying on the typing module's definition.